+++ pkg/papi/active_property_hostname_test.go
@@ -160,6 +160,269 @@ func TestPapiListActivePropertyHostnames(t *testing.T) {
 				},
 			},
 		},
+		"200 OK - all types": {
+			params: ListActivePropertyHostnamesRequest{
+				PropertyID:        "prp_175780",
+				IncludeCertStatus: true,
+			},
+			responseStatus: http.StatusOK,
+			responseBody: `
+				{
+				  "accountId": "act_123",
+				  "availableSort": [
+					"hostname:a",
+					"hostname:d"
+				  ],
+				  "contractId": "ctr_123",
+				  "currentSort": "hostname:a",
+				  "defaultSort": "hostname:a",
+				  "groupId": "grp_15225",
+				  "hostnames": {
+					"currentItemCount": 5,
+					"items": [
+					  {
+						"cnameFrom": "example.com",
+						"cnameType": "EDGE_HOSTNAME",
+						"productionCertType": "DEFAULT",
+						"productionCnameTo": "example.com.edgekey.net",
+						"productionEdgeHostnameId": "ehn_895822"
+					  },
+					  {
+						"cnameFrom": "m-example.com",
+						"cnameType": "EDGE_HOSTNAME",
+						"stagingCertType": "DEFAULT",
+						"stagingCnameTo": "m-example.com.edgekey.net",
+						"stagingEdgeHostnameId": "ehn_293412"
+					  },
+					  {
+						"certStatus": {
+						  "authorization": {
+							"dns01": {
+							  "result": {
+								"message": "dns01 cps dry run cname/TXT incomplete",
+								"src": "CPS",
+								"timestamp": "2024-07-25T16:17:37Z"
+							  },
+							  "value": "dummy-unique-value-for-DNS-TXT-record"
+							},
+							"http01": {
+							  "body": "unique http body content",
+							  "result": {
+								"message": "http01 cps dry run fail reason",
+								"src": "CPS",
+								"timestamp": "2024-07-25T16:17:37Z"
+							  },
+							  "url": "/.well-known/acme-challenge/"
+							},
+							"status": "ATTEMPTING_VALIDATION",
+							"validUntil": "2024-07-25T16:17:37Z"
+						  },
+						  "certExpirationDate": "2024-07-25T16:17:37Z",
+						  "production": [
+							{
+							  "status": "PENDING"
+							}
+						  ],
+						  "staging": [
+							{
+							  "status": "PENDING"
+							}
+						  ],
+						  "validationCname": {
+							"hostname": "_acme-challenge.www.example.com",
+							"target": "{token}.www.example.com.akamai-domain.com"
+						  }
+						},
+						"cnameFrom": "example2.com",
+						"cnameType": "EDGE_HOSTNAME",
+						"stagingCertType": "DEFAULT",
+						"stagingCnameTo": "example2.com.edgekey.net",
+						"stagingEdgeHostnameId": "ehn_895822"
+					  },
+					  {
+						"ccmCertStatus": {
+						  "ecdsaStagingStatus": "DEPLOYED",
+						  "rsaStagingStatus": "DEPLOYED"
+						},
+						"ccmCertificates": {
+						  "ecdsaCertId": "98765",
+						  "ecdsaCertLink": "/ccm/v1/certificates/98765",
+						  "rsaCertId": "12345",
+						  "rsaCertLink": "/ccm/v1/certificates/12345"
+						},
+						"cnameFrom": "www.example-ccm.com",
+						"cnameType": "EDGE_HOSTNAME",
+						"mtls": {
+						  "caSetId": "524125",
+						  "caSetLink": "/mtls-edge-truststore/v2/ca-sets/524125",
+						  "checkClientOcsp": false,
+						  "sendCaSetClient": false
+						},
+						"stagingCertType": "CCM",
+						"stagingEdgeHostnameId": "ehn_7123",
+						"tlsConfiguration": {
+						  "cipherProfile": "ak-akamai-2020q1",
+						  "disallowedTlsVersions": [
+							"TLSv1_1",
+							"TLSv1"
+						  ],
+						  "fipsMode": false,
+						  "stapleServerOcspResponse": true
+						}
+					  },
+					  {
+						"ccmCertStatus": {
+						  "ecdsaProductionStatus": "DEPLOYED",
+						  "rsaProductionStatus": "DEPLOYED"
+						},
+						"ccmCertificates": {
+						  "ecdsaCertId": "98765",
+						  "ecdsaCertLink": "/ccm/v1/certificates/98765",
+						  "rsaCertId": "12345",
+						  "rsaCertLink": "/ccm/v1/certificates/12345"
+						},
+						"cnameFrom": "www.example-ccm.com",
+						"cnameType": "EDGE_HOSTNAME",
+						"mtls": {
+						  "caSetId": "524125",
+						  "caSetLink": "/mtls-edge-truststore/v2/ca-sets/524125",
+						  "checkClientOcsp": false,
+						  "sendCaSetClient": false
+						},
+						"productionCertType": "CCM",
+						"productionEdgeHostnameId": "ehn_7123",
+						"tlsConfiguration": {
+						  "cipherProfile": "ak-akamai-2020q1",
+						  "disallowedTlsVersions": [
+							"TLSv1_1",
+							"TLSv1"
+						  ],
+						  "fipsMode": false,
+						  "stapleServerOcspResponse": true
+						}
+					  }
+					],
+					"nextLink": "/papi/v1/properties/prp_175780/hostnames?offset=1&groupId=grp_15225&contractId=ctr_K-0N7RAK71&limit=3",
+					"totalItems": 5
+				  },
+				  "propertyId": "prp_175780",
+				  "propertyName": "mytestproperty.com"
+				}`,
+			expectedPath: "/papi/v1/properties/prp_175780/hostnames?includeCertStatus=true",
+			expectedResponse: &ListActivePropertyHostnamesResponse{
+				AccountID:    "act_123",
+				ContractID:   "ctr_123",
+				GroupID:      "grp_15225",
+				PropertyID:   "prp_175780",
+				PropertyName: "mytestproperty.com",
+				DefaultSort:  SortAscending,
+				CurrentSort:  SortAscending,
+				AvailableSort: []SortOrder{
+					SortAscending,
+					SortDescending,
+				},
+				Hostnames: HostnamesResponseItems{
+					CurrentItemCount: 5,
+					TotalItems:       5,
+					Items: []HostnameItem{
+						{
+							CnameFrom:                "example.com",
+							CnameType:                HostnameCnameTypeEdgeHostname,
+							ProductionCertType:       CertTypeDefault,
+							ProductionCnameTo:        "example.com.edgekey.net",
+							ProductionEdgeHostnameID: "ehn_895822",
+						},
+						{
+							CnameFrom:             "m-example.com",
+							CnameType:             HostnameCnameTypeEdgeHostname,
+							StagingCertType:       CertTypeDefault,
+							StagingEdgeHostnameID: "ehn_293412",
+							StagingCnameTo:        "m-example.com.edgekey.net",
+						},
+						{
+							CnameFrom:             "example2.com",
+							CnameType:             HostnameCnameTypeEdgeHostname,
+							StagingCertType:       CertTypeDefault,
+							StagingEdgeHostnameID: "ehn_895822",
+							StagingCnameTo:        "example2.com.edgekey.net",
+							CertStatus: &CertStatusItem{
+								Production: []StatusItem{
+									{
+										Status: "PENDING",
+									},
+								},
+								Staging: []StatusItem{
+									{
+										Status: "PENDING",
+									},
+								},
+								ValidationCname: ValidationCname{
+									Hostname: "_acme-challenge.www.example.com",
+									Target:   "{token}.www.example.com.akamai-domain.com",
+								},
+							},
+						},
+						{
+							CnameFrom:             "www.example-ccm.com",
+							CnameType:             HostnameCnameTypeEdgeHostname,
+							StagingCertType:       CertTypeCCM,
+							StagingEdgeHostnameID: "ehn_7123",
+							MTLS: &MTLS{
+								CASetID:         "524125",
+								CASetLink:       "/mtls-edge-truststore/v2/ca-sets/524125",
+								CheckClientOCSP: false,
+								SendCASetClient: false,
+							},
+							TLSConfiguration: &TLSConfiguration{
+								CipherProfile:            "ak-akamai-2020q1",
+								DisallowedTLSVersions:    []string{"TLSv1_1", "TLSv1"},
+								FIPSMode:                 false,
+								StapleServerOcspResponse: true, //TODO OCSP
+							},
+							CCMCertificates: &CCMCertificates{
+								ECDSACertID:   "98765",
+								ECDSACertLink: "/ccm/v1/certificates/98765",
+								RSACertID:     "12345",
+								RSACertLink:   "/ccm/v1/certificates/12345",
+							},
+							CCMCertStatus: &CCMCertStatus{
+								ECDSAStagingStatus: "DEPLOYED",
+								RSAStagingStatus:   "DEPLOYED",
+							},
+						},
+						{
+							CnameFrom:                "www.example-ccm.com",
+							CnameType:                HostnameCnameTypeEdgeHostname,
+							ProductionCertType:       CertTypeCCM,
+							ProductionEdgeHostnameID: "ehn_7123",
+							MTLS: &MTLS{
+								CASetID:         "524125",
+								CASetLink:       "/mtls-edge-truststore/v2/ca-sets/524125",
+								CheckClientOCSP: false,
+								SendCASetClient: false,
+							},
+							TLSConfiguration: &TLSConfiguration{
+								CipherProfile:            "ak-akamai-2020q1",
+								DisallowedTLSVersions:    []string{"TLSv1_1", "TLSv1"},
+								FIPSMode:                 false,
+								StapleServerOcspResponse: true,
+							},
+							CCMCertificates: &CCMCertificates{
+								ECDSACertID:   "98765",
+								ECDSACertLink: "/ccm/v1/certificates/98765",
+								RSACertID:     "12345",
+								RSACertLink:   "/ccm/v1/certificates/12345",
+							},
+							CCMCertStatus: &CCMCertStatus{
+								ECDSAProductionStatus: "DEPLOYED",
+								RSAProductionStatus:   "DEPLOYED",
+							},
+						},
+					},
+					NextLink: ptr.To("/papi/v1/properties/prp_175780/hostnames?offset=1&groupId=grp_15225&contractId=ctr_K-0N7RAK71&limit=3"),
+				},
+			},
+		},
 		"validation error PropertyID missing": {
 			params: ListActivePropertyHostnamesRequest{
 				Offset: 3,